Move all drawing to element
